+++ src/main/java/tornadofx/ViewModelFacades.kt
@@ -4,6 +4,7 @@ import javafx.beans.Observable
 import javafx.beans.property.*
 import javafx.beans.value.ObservableValue
 import javafx.collections.ObservableList
+import javafx.collections.ObservableMap
 import javafx.collections.ObservableSet
 
 interface BindingAware
@@ -93,6 +94,32 @@ class BindingAwareSimpleSetProperty<T>(viewModel: ViewModel, name: String?) : Si
     }
 
     override fun bindContentBidirectional(list: ObservableSet<T>?) {
+        super.bindContentBidirectional(list)
+        recordBinding(list)
+    }
+
+    fun recordBinding(observableValue: Observable?) {
+        if (observableValue != null) {
+            ViewModel.propertyToFacade[observableValue] = this
+            ViewModel.propertyToViewModel[observableValue] = bean as ViewModel
+        }
+    }
+
+    /**
+     * Return a unique id for this object instead of the default hashCode which is dependent on the children.
+     * Without this override we wouldn't be able to identify the facade in our internal maps.
+     */
+    override fun hashCode() = System.identityHashCode(this)
+    override fun equals(other: Any?) = this === other
+}
+
+class BindingAwareSimpleMapProperty<S, T>(viewModel: ViewModel, name: String?) : SimpleMapProperty<S, T>(viewModel, name) {
+    override fun bind(newObservable: ObservableValue<out ObservableMap<S, T>>?) {
+        super.bind(newObservable)
+        recordBinding(newObservable)
+    }
+
+    override fun bindContentBidirectional(list: ObservableMap<S, T>?) {
         super.bindContentBidirectional(list)
         recordBinding(list)
     }